lab_5/pyproject.toml
[project]
name = "lab-5"
version = "0.1.0"
description = "Vector Quantization using LBG"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "numpy>=2.0.0",
    "pillow>=10.0.0",
]

[dependency-groups]
tests = [
    "pytest>=8.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]